// resources/js/types/Question.d.ts
export interface Question {
    id:                 string;
    subtest_id:         string;
    subject_id:         string;
    option_a:           string;
    option_b:           string;
    option_c:           string;
    option_d:           string;
    option_e:           string;
    correct_answer:     string;
    explanation:        string;
}